Finished_Notebook5_Planned06+07
I finished running through Notebook 05, the changes I made from the boilerplate code are heavy in the end because it was extremely obvious from the simplest metrics that the random model wasn't going to work. It had to be tested, but I saw no reason in polishing the end of the report with visualisations. I have a plan for 06 + 07 now, 06 is going to be quite large but given the time constraints I didn't want to create two work books. Once done I was hoping you could have a crack at 07 as it's definitely in your wheelhouse and the code you've written will fit. What's there now are just placeholder titles, we'll see how 06 goes :-)